jokes_main/views.py
from rest_framework.decorators import api_view
from rest_framework.response import Response

@api_view()
def root_route(request):
    return Response({
        "message": "This is an API for 'Jokes All Day'"
    })